requirements.txt
pyTelegramBotAPI==4.12.0
Pillow==10.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0